+++ quam_builder/architecture/superconducting/components/readout_resonator.py
@@ -1,4 +1,5 @@
-from typing import Optional
+from typing import Optional, Dict, Any
+from dataclasses import field
 
 from quam.core import quam_dataclass
 from quam.components.channels import InOutIQChannel, InOutMWChannel
@@ -44,10 +45,10 @@ class ReadoutResonatorBase:
     
     kappa: float = 1000000
 
+    extras: Dict[str, Any] = field(default_factory=dict)
+
     @staticmethod
-    def calculate_voltage_scaling_factor(
-        fixed_power_dBm: float, target_power_dBm: float
-    ):
+    def calculate_voltage_scaling_factor(fixed_power_dBm: float, target_power_dBm: float):
         """
         Calculate the voltage scaling factor required to scale fixed power to target power.
 
@@ -111,9 +112,7 @@ class ReadoutResonatorIQ(InOutIQChannel, ReadoutResonatorBase):
             ValueError: If `gain` or `amplitude` is outside their valid ranges.
 
         """
-        return set_output_power_iq_channel(
-            self, power_in_dbm, gain, max_amplitude, Z, operation
-        )
+        return set_output_power_iq_channel(self, power_in_dbm, gain, max_amplitude, Z, operation)
 
 
 @quam_dataclass